requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.28.55
botocore==1.31.85
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
dnspython==2.7.0
dulwich==0.21.7
fastapi==0.115.4
fastjsonschema==2.20.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
jmespath==1.0.1
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
pexpect==4.9.0
pika==1.3.2
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.6.3
PyMySQL==1.1.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
rococo==1.0.23
s3transfer==0.7.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
surrealdb==0.3.2
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.0.7
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.0
watchfiles==0.24.0
websockets==10.4
